Fix keyhandling in the classic greeter
As we needed public exposure to the setFocus() slot that we don't have from plasma but do have in PasswordLineEdit
This was changed to be more generic, and now simply a better version of Plasma::LineEdit but with a property for setting echo mode

All themes were then updated.

This change becomes irrelvant when we move to PlasmaComponents that has a line edit with proper focus and echo mode handling.